A little spring cleaning
Code formatting and org
