+++ src/main/scala/org/espercatz/domain/AssociationRule.scala
@@ -1,0 +1,8 @@
+package org.espercatz.domain
+
+case class AssociationRule(
+  antecedent: Set[Int],
+  consequent: Set[Int],
+  confidence: Double,
+  support: Int
+)
