+++ src/main/scala/org/espercatz/domain/ItemSet.scala
@@ -1,0 +1,3 @@
+package org.espercatz.domain
+
+case class ItemSet(items: Set[Int], size: Int, support: Int)
